// prisma/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AuthProvider = {
  CREDENTIAL: 'CREDENTIAL',
  GOOGLE: 'GOOGLE'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BLOCK: 'BLOCK'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const Role = {
  STUDENT: 'STUDENT',
  EXPERT: 'EXPERT',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const AssignmentStatus = {
  OPEN: 'OPEN',
  ASSIGNED: 'ASSIGNED',
  AWAITING_PAYMENT: 'AWAITING_PAYMENT',
  IN_PROGRESS: 'IN_PROGRESS',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  DISPUTED: 'DISPUTED'
} as const

export type AssignmentStatus = (typeof AssignmentStatus)[keyof typeof AssignmentStatus]


export const BidStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type BidStatus = (typeof BidStatus)[keyof typeof BidStatus]


export const PaymentStatus = {
  INITIATED: 'INITIATED',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const EscrowStatus = {
  HELD: 'HELD',
  RELEASED_TO_EXPERT: 'RELEASED_TO_EXPERT',
  REFUNDED_TO_STUDENT: 'REFUNDED_TO_STUDENT'
} as const

export type EscrowStatus = (typeof EscrowStatus)[keyof typeof EscrowStatus]


export const ExpertVerificationStatus = {
  PENDING: 'PENDING',
  APPROVE: 'APPROVE',
  REJECT: 'REJECT'
} as const

export type ExpertVerificationStatus = (typeof ExpertVerificationStatus)[keyof typeof ExpertVerificationStatus]


export const PaymentGateway = {
  BKASH: 'BKASH',
  SSLCOMMERZ: 'SSLCOMMERZ',
  STRIPE: 'STRIPE'
} as const

export type PaymentGateway = (typeof PaymentGateway)[keyof typeof PaymentGateway]